refactor: keep inference helpers in inference.rs for cohesion
Move assemble_context(), collect_stream(), try_with_rate_limit(),
try_overflow_recovery(), and preflight_compact_if_needed() back into
inference.rs — they are implementation details of the inference loop,
not reusable utilities. Reading one flow shouldn't require jumping
between two files.

inference_helpers.rs reverts to its original scope: generic utilities
(estimate_tokens, error classifiers, backoff calculation, constants).

inference.rs: 781 lines (cohesive unit > arbitrary line cap)
inference_helpers.rs: 161 lines (generic utilities only)
