+++ koda-core/src/inference_helpers.rs
@@ -1,10 +1,7 @@
 //! Helper functions for inference — context estimation, message assembly,
 //! error classification.
 
-use crate::persistence::Persistence;
 use crate::providers::{ChatMessage, ToolCall};
-use tokio::sync::mpsc;
-use tokio_util::sync::CancellationToken;
 
 /// Pre-flight context budget threshold (percentage).
 /// If context usage exceeds this before calling the provider, auto-compact first.
@@ -40,7 +37,7 @@ pub fn estimate_tokens(messages: &[ChatMessage]) -> usize {
 }
 
 /// Assemble messages from DB history into ChatMessage vec.
-fn assemble_messages(
+pub fn assemble_messages(
     system_message: &ChatMessage,
     history: &[crate::db::Message],
 ) -> Vec<ChatMessage> {
@@ -96,376 +93,6 @@ pub fn is_context_overflow_error(err: &anyhow::Error) -> bool {
         || msg.contains("exceeds the model")
         || msg.contains("request too large")
         || (msg.contains("413") && msg.contains("too large"))
-}
-
-/// Pre-flight budget check: if context usage exceeds the threshold, compact
-/// before sending to the provider. Re-assembles context after successful compaction.
-///
-/// Returns the (possibly updated) message vec.
-pub async fn preflight_compact_if_needed(
-    messages: Vec<ChatMessage>,
-    db: &crate::db::Database,
-    session_id: &str,
-    system_message: &ChatMessage,
-    pending_images: Option<&[crate::providers::ImageData]>,
-    iteration: u32,
-    config: &crate::config::KodaConfig,
-    provider: &dyn crate::providers::LlmProvider,
-    sink: &dyn crate::engine::EngineSink,
-) -> anyhow::Result<Vec<ChatMessage>> {
-    let ctx_pct = crate::context::percentage();
-    if ctx_pct < PREFLIGHT_COMPACT_THRESHOLD {
-        return Ok(messages);
-    }
-
-    tracing::warn!("Pre-flight: context at {ctx_pct}%, attempting auto-compact");
-    sink.emit(crate::engine::EngineEvent::Info {
-        message: format!("\u{1f4e6} Context at {ctx_pct}% \u{2014} compacting before sending..."),
-    });
-
-    match crate::compact::compact_session_with_provider(
-        db,
-        session_id,
-        config.max_context_tokens,
-        &config.model_settings,
-        provider,
-    )
-    .await
-    {
-        Ok(Ok(result)) => {
-            sink.emit(crate::engine::EngineEvent::Info {
-                message: format!(
-                    "\u{2705} Compacted {} messages (~{} token summary)",
-                    result.deleted, result.summary_tokens
-                ),
-            });
-            // Re-assemble with compacted history
-            assemble_context(
-                db,
-                session_id,
-                system_message,
-                pending_images,
-                iteration,
-                config.max_context_tokens,
-            )
-            .await
-        }
-        Ok(Err(skip)) => {
-            tracing::info!("Pre-flight compact skipped: {skip:?}");
-            if matches!(skip, crate::compact::CompactSkip::HistoryTooLarge) {
-                sink.emit(crate::engine::EngineEvent::Warn {
-                    message: "\u{26a0}\u{fe0f} Context is full but history is too large for this model to summarize. \
-                              Start a new session (/session) or switch to a model with a larger context window."
-                        .to_string(),
-                });
-            }
-            Ok(messages)
-        }
-        Err(e) => {
-            tracing::warn!("Pre-flight compact failed: {e:#}");
-            sink.emit(crate::engine::EngineEvent::Warn {
-                message: format!("Compact failed: {e:#}. Continuing anyway..."),
-            });
-            Ok(messages)
-        }
-    }
-}
-
-/// Recover from a context overflow error: compact the session, re-assemble
-/// context, and retry the provider call once.
-///
-/// Returns `Ok(Some((rx, messages)))` on success (receiver + updated messages),
-/// `Ok(None)` if cancelled during retry, or `Err` if compaction/retry fails.
-pub async fn try_overflow_recovery(
-    original_err: anyhow::Error,
-    db: &crate::db::Database,
-    session_id: &str,
-    system_message: &ChatMessage,
-    pending_images: Option<&[crate::providers::ImageData]>,
-    iteration: u32,
-    config: &crate::config::KodaConfig,
-    provider: &dyn crate::providers::LlmProvider,
-    tool_defs: &[crate::providers::ToolDefinition],
-    cancel: &CancellationToken,
-    sink: &dyn crate::engine::EngineSink,
-) -> Result<
-    Option<(
-        mpsc::Receiver<crate::providers::StreamChunk>,
-        Vec<ChatMessage>,
-    )>,
-    anyhow::Error,
-> {
-    use anyhow::Context;
-
-    sink.emit(crate::engine::EngineEvent::SpinnerStop);
-    sink.emit(crate::engine::EngineEvent::Warn {
-        message:
-            "\u{26a0}\u{fe0f} Provider rejected request (context overflow). Compacting and retrying..."
-                .to_string(),
-    });
-    tracing::warn!("Context overflow from provider: {original_err:#}");
-
-    // Try to compact
-    match crate::compact::compact_session_with_provider(
-        db,
-        session_id,
-        config.max_context_tokens,
-        &config.model_settings,
-        provider,
-    )
-    .await
-    {
-        Ok(Ok(result)) => {
-            sink.emit(crate::engine::EngineEvent::Info {
-                message: format!(
-                    "\u{2705} Compacted {} messages. Retrying...",
-                    result.deleted
-                ),
-            });
-        }
-        _ => {
-            return Err(original_err)
-                .context("LLM inference failed (context overflow, compaction unsuccessful)");
-        }
-    }
-
-    // Re-assemble messages with compacted history
-    let messages = assemble_context(
-        db,
-        session_id,
-        system_message,
-        pending_images,
-        iteration,
-        config.max_context_tokens,
-    )
-    .await?;
-
-    // Retry once
-    sink.emit(crate::engine::EngineEvent::SpinnerStart {
-        message: "Retrying...".into(),
-    });
-    let rx = tokio::select! {
-        result = provider.chat_stream(&messages, tool_defs, &config.model_settings) => {
-            result.context("LLM inference failed after compaction retry")?
-        }
-        _ = cancel.cancelled() => return Ok(None),
-    };
-    Ok(Some((rx, messages)))
-}
-
-/// Attempt to start a chat stream with exponential backoff on rate limits.
-///
-/// Returns `Ok(Some(rx))` on success, `Ok(None)` if cancelled during retries,
-/// or `Err` for non-retriable failures.
-pub async fn try_with_rate_limit(
-    provider: &dyn crate::providers::LlmProvider,
-    messages: &[ChatMessage],
-    tool_defs: &[crate::providers::ToolDefinition],
-    model_settings: &crate::config::ModelSettings,
-    cancel: &CancellationToken,
-    sink: &dyn crate::engine::EngineSink,
-) -> Result<Option<mpsc::Receiver<crate::providers::StreamChunk>>, anyhow::Error> {
-    let mut last_err = None;
-    for attempt in 0..RATE_LIMIT_MAX_RETRIES {
-        let result = tokio::select! {
-            result = provider.chat_stream(messages, tool_defs, model_settings) => result,
-            _ = cancel.cancelled() => return Ok(None),
-        };
-        match result {
-            Ok(rx) => return Ok(Some(rx)),
-            Err(e) if is_rate_limit_error(&e) && attempt + 1 < RATE_LIMIT_MAX_RETRIES => {
-                let delay = rate_limit_backoff(attempt);
-                sink.emit(crate::engine::EngineEvent::SpinnerStop);
-                sink.emit(crate::engine::EngineEvent::Warn {
-                    message: format!("\u{23f3} Rate limited. Retrying in {}s...", delay.as_secs()),
-                });
-                tracing::warn!(
-                    "Rate limit (attempt {}/{}): {e:#}",
-                    attempt + 1,
-                    RATE_LIMIT_MAX_RETRIES
-                );
-                tokio::time::sleep(delay).await;
-                sink.emit(crate::engine::EngineEvent::SpinnerStart {
-                    message: format!("Retrying (attempt {})...", attempt + 2),
-                });
-                last_err = Some(e);
-            }
-            Err(e) => return Err(e),
-        }
-    }
-    Err(last_err.unwrap_or_else(|| anyhow::anyhow!("Rate limit retries exhausted")))
-}
-
-/// Result of collecting a streamed LLM response.
-pub struct StreamResult {
-    /// Accumulated text content from the response.
-    pub text: String,
-    /// Tool calls requested by the model.
-    pub tool_calls: Vec<ToolCall>,
-    /// Token usage statistics.
-    pub usage: crate::providers::TokenUsage,
-    /// Total character count of text deltas.
-    pub char_count: usize,
-    /// Whether the stream was interrupted by cancellation.
-    pub interrupted: bool,
-}
-
-/// Collect a streamed LLM response, emitting engine events for thinking/text/tool calls.
-///
-/// Handles thinking ↔ response state transitions, cancellation via `CancellationToken`,
-/// and spinner lifecycle. Returns a `StreamResult` — the caller is responsible for
-/// persistence and early-return on interruption.
-pub async fn collect_stream(
-    rx: &mut mpsc::Receiver<crate::providers::StreamChunk>,
-    sink: &dyn crate::engine::EngineSink,
-    cancel: &CancellationToken,
-) -> StreamResult {
-    use crate::providers::StreamChunk;
-
-    let mut full_text = String::new();
-    let mut tool_calls: Vec<ToolCall> = Vec::new();
-    let mut usage = crate::providers::TokenUsage::default();
-    let mut first_token = true;
-    let mut char_count: usize = 0;
-    let mut native_think_buf = String::new();
-    let mut response_banner_shown = false;
-    let mut thinking_banner_shown = false;
-    let mut interrupted = false;
-
-    loop {
-        let chunk = tokio::select! {
-            c = rx.recv() => c,
-            _ = cancel.cancelled() => {
-                interrupted = true;
-                None
-            }
-        };
-
-        if interrupted || cancel.is_cancelled() {
-            sink.emit(crate::engine::EngineEvent::SpinnerStop);
-            if !full_text.is_empty() {
-                sink.emit(crate::engine::EngineEvent::TextDone);
-            }
-            sink.emit(crate::engine::EngineEvent::Warn {
-                message: "Interrupted".into(),
-            });
-            return StreamResult {
-                text: full_text,
-                tool_calls,
-                usage,
-                char_count,
-                interrupted: true,
-            };
-        }
-
-        let Some(chunk) = chunk else { break };
-
-        match chunk {
-            StreamChunk::TextDelta(delta) => {
-                if first_token {
-                    // Close any open thinking block (content already streamed)
-                    if !native_think_buf.is_empty() {
-                        sink.emit(crate::engine::EngineEvent::SpinnerStop);
-                        sink.emit(crate::engine::EngineEvent::ThinkingDone);
-                        native_think_buf.clear();
-                        thinking_banner_shown = true;
-                    }
-                    sink.emit(crate::engine::EngineEvent::SpinnerStop);
-                    first_token = false;
-                }
-
-                // Show response banner on first non-empty text
-                if !response_banner_shown && !delta.trim().is_empty() {
-                    sink.emit(crate::engine::EngineEvent::ResponseStart);
-                    response_banner_shown = true;
-                }
-
-                full_text.push_str(&delta);
-                char_count += delta.len();
-                sink.emit(crate::engine::EngineEvent::TextDelta {
-                    text: delta.clone(),
-                });
-            }
-            StreamChunk::ThinkingDelta(delta) => {
-                if !thinking_banner_shown {
-                    sink.emit(crate::engine::EngineEvent::SpinnerStop);
-                    sink.emit(crate::engine::EngineEvent::ThinkingStart);
-                    thinking_banner_shown = true;
-                }
-                sink.emit(crate::engine::EngineEvent::ThinkingDelta {
-                    text: delta.clone(),
-                });
-                native_think_buf.push_str(&delta);
-            }
-            StreamChunk::ToolCalls(tcs) => {
-                if !native_think_buf.is_empty() {
-                    sink.emit(crate::engine::EngineEvent::SpinnerStop);
-                    sink.emit(crate::engine::EngineEvent::ThinkingDone);
-                    native_think_buf.clear();
-                }
-                sink.emit(crate::engine::EngineEvent::SpinnerStop);
-                tool_calls = tcs;
-            }
-            StreamChunk::Done(u) => {
-                if !native_think_buf.is_empty() {
-                    sink.emit(crate::engine::EngineEvent::SpinnerStop);
-                    sink.emit(crate::engine::EngineEvent::ThinkingDone);
-                    native_think_buf.clear();
-                }
-                usage = u;
-                break;
-            }
-        }
-    }
-
-    // Flush remaining text
-    sink.emit(crate::engine::EngineEvent::TextDone);
-
-    if first_token {
-        sink.emit(crate::engine::EngineEvent::SpinnerStop);
-    }
-
-    StreamResult {
-        text: full_text,
-        tool_calls,
-        usage,
-        char_count,
-        interrupted: false,
-    }
-}
-
-/// Load conversation history, assemble messages with the system prompt,
-/// attach pending images (first iteration only), and update context tracking.
-///
-/// This is the single source of truth for context assembly — called on initial
-/// build, after pre-flight compaction, and after overflow recovery.
-pub async fn assemble_context(
-    db: &crate::db::Database,
-    session_id: &str,
-    system_message: &ChatMessage,
-    pending_images: Option<&[crate::providers::ImageData]>,
-    iteration: u32,
-    max_context_tokens: usize,
-) -> anyhow::Result<Vec<ChatMessage>> {
-    let history = db.load_context(session_id).await?;
-    let mut messages = assemble_messages(system_message, &history);
-
-    // Attach pending images to the last user message (first iteration only)
-    if iteration == 0 {
-        if let Some(imgs) = pending_images {
-            if !imgs.is_empty() {
-                if let Some(last_user) = messages.iter_mut().rev().find(|m| m.role == "user") {
-                    last_user.images = Some(imgs.to_vec());
-                }
-            }
-        }
-    }
-
-    // Track context window usage
-    let context_used = estimate_tokens(&messages);
-    crate::context::update(context_used, max_context_tokens);
-
-    Ok(messages)
 }
 
 #[cfg(test)]
